add decrypt.rs for decryption of general data regardless of key derivation method
