+++ mm2src/crypto/src/encrypt.rs
@@ -65,6 +65,26 @@ pub struct EncryptedData {
     pub tag: String,
 }
 
+/// Encrypts the provided data using AES-256-CBC encryption and HMAC for integrity check.
+///
+/// This function performs several operations:
+/// - It generates an Initialization Vector (IV) for the AES encryption.
+/// - It creates an AES-256-CBC cipher instance and encrypts the data with the provided key and the generated IV.
+/// - It creates an HMAC tag for verifying the integrity of the encrypted data.
+/// - It constructs an [`EncryptedData`] instance containing all the necessary components for decryption.
+///
+/// # Arguments
+/// * `data` - A byte slice reference to the data that needs to be encrypted.
+/// * `key_derivation_details` - A [`KeyDerivationDetails`] instance containing detailed information about the key derivation process.
+/// * `key_aes` - A byte array reference to the AES key used for encryption.
+/// * `key_hmac` - A byte array reference to the HMAC key used for creating the HMAC tag.
+///
+/// # Returns
+/// `MmResult<EncryptedData, EncryptionError>` - The result is either an [`EncryptedData`]
+/// struct containing all the necessary components for decryption, or an [`EncryptionError`] in case of failure.
+///
+/// # Errors
+/// This function can return various errors related to IV generation, AES encryption, HMAC creation, and data encoding.
 pub fn encrypt_data(
     data: &[u8],
     key_derivation_details: KeyDerivationDetails,
